Week check on menus
Since not all menues are update on time (before 1030 on mondays) a check
has been added. If the correct week cannot be found a 'No menu for week
X' is printed instead.

A rework was done on where the restaurant name was handled. Before, the
day function returned the name. But now there is a special function
(called namefromurl) since the week check is done outside the day
function. The reason was that else there would be the same week check
logic in all day functions.

An attempt to split the days in separate tables was made (hence the
table width 55em in the stylesheet), but the first column got different
widths eventhough it always contains the same entries for all tables.