Reworked tables
Instead of having one large table, there is now one table for each day.
This opens for using <h2> for name of day, and also having a <a> tag id
for each day. So now it is possible to refer to a specific day by adding
hash-dayname (#torsdag for instance).
<tbody> tag was re-added. Let's see what validator.org says about that.