Minor stylesheet adjustments.
The text was dangerously close to the table edge. The stylesheet was
modified to keep a 0.5em left and 1em right margin (padding).

The deploy/reploy duo was changed to echo cp info instead of keeping it
in a variable and then printing. However, now an empty line is printed
if the file is not updated.