Better handling when link to menu is broken.
Added 'no workie' text like when no menu for current week can be found
when link is broken.

Link to Scotland (Gotland) Yard was moved one id to 114576. Hope this is
not a recurring event.

Finn Inn now have the word 'vecka' and the actual weeknumber on separate
lines. The solution is to replace all newlines '\n' with a space before
parsing the different weekdays (for all restaurants). This acually is a
nicer solution than before.