Fixed Scotland Yard (again)
The moving of menu for Gotland Yard from one id to another was not a
isolated event. It happened again. So now we use a link which points to
the page 'above' the menu. The menu is also present on this page, but no
week info is present. The comical thing is there is an option to save
the link to the menu page. Which will be invalid every week. Go figure.

Since there is no week info anylonger the check was reworked. Now there
is another function associated with every url. That is, there is a list
of functions: one handles a day (as before) and another checks if we
have the correct menu for this week. In the case of Scotland Yard, the
week check func makes no check but returns 1 always. Nice. Took awile to
figure out the syntax for calling one of many functions in a list stored
in a hash: $hash{$key}[n]->($param1, $param2,...);