Release 1.3.2 Fixed Ideon Alfa
For some time Ideon Alfa has been misbehaving. It used to work by
looking for <h2> or <h3> as a day separator. However this is no longer
true between onsdag and torsdag. So the onsdag also included torsdag,
torsdag had fredags menu and fredag was correct. Now instead we use the
<span preceeding the day and we get everything up to Vegetarisk and to
the next </p>. Seems to work. We have to see for how long.