Jump to today by javascript.
Small javascript added which jumps to today in menu. If today is
the weekend, no jump happens. Also, for monday there is no jump, because
it is useful to see the h1 header to check that the page has updated for
the week. It might be necessary to add date on each day in the future.

Also, using the perl quoting operator qq makes the html code much more
readable. For now only the header and footer uses qq.