Better way to jump to today
Using location.hash to set today also changed the URL, which is not
optimal. Firstly, it looks bad. But it is bad for other reasons.
For instance, if a bookmark is set on a 'onsdag', then the
default day for that person would be 'onsdag'. Since we do not want
to scroll on mondays since then we would not see the week dates, we
use the default day. Which will then be 'onsdag'. Using scrollIntoView
does not change the URL.