Jump to today revisited.
It seems like scrollIntoView at the end of the page does not work
in Android. It is not a problem with scrollIntoView however. It seems
the onload event never fires. The javascript has been moved to the <head>
and put in a proper function. It is called from window.onload which works
in firefox at least. By clicking the main header, you can also jump to today.

A lonely ampersand from delta made us not go through the validator. Fixed.