Added meta name viewport.
To mitigate some problems with Chrome mobile (th renders very small)
a meta tag has been added, viewport. This makes all text equal in size.
The drawback is that a menu item no longer fits on one text line. It
also seems that it is not possible to zoom out. What ever...