Menus in sorted order.
Previously, the url (as hashed) was used to get the order for the
different restaurants (default perl behaviour). Now a specific sort
function is used which uses the name of the restaurant as sorting key.

Also, LAGK has been removed (temporarily) since there menu has moved to
some generic place called catera.se, and the URL no longer works as is
(the new URL does also say 500 as error code).