Fixed Sarminer missing one item sometimes
If the colon (:) on Dagens or Vegetarisk came after the bold <b> tag
instead of before, a specific cleaning regex removed the colon which
lead to the Dagens/Vegetarisk cleaning regex removed the entire item.
Not sure why the colon cleaner was put in in the first place. But it is
not needed anymore.