Fixed Scotland Yard
The <strong> tag is now much misused with many empty strong entries. So
this is no longer a good end marker of a day. An end table tag was
usable except for Monday. By adding an empty <td> item before the end
table tag in the regex, all days were matched.