Fixed new Finn Inn layout
Finn Inn is remaking it's web page. There is no current menu though, so
to be able to test with the menu from w 51 a new option (-w) has been
implemented to override the current week number. The dates for the week
are not correct however (uses the dates for the current week), so this
is a debug option only.