Matsalen off by one fixed
Matsalen had torsdag menu on wednesday, etc. The tag between day and
menu items was changed from </p> to <br />. Both are now accepted. It is
high time changes in layout is parsed by adding to the rules instead of
changing it. Changing the rule will often break the parsing when the
layout is reverted.

Also, a lot of nbsp were present in the menu for scotland yard. These
are normally turned into soft spaces, but they were coded as &amp;&nbsp;
which was not handled. It is now.