Fixed ideon delta url mapping
Restaurant Ideon Delta have a page which contains the proper url for a
specific week. The proper url is hard (impossible) to calculate, so the
mapping page needs to be downloaded and the proper url found on that
page.

Also, since curl is used twice get webpages for url, a subroutine was
created which returns req status and web page body for a url.