Fixed finn inn and one other
The finn inn menu page contained som 'alternative' encoded space which
needed to be converted to normal spaces. Anothe menu page (which has
been forgotten since) had no spave between the week marker and the week
number so now 0-3 spaces are allowed after the marker and an optional .
(which can be anything).