Fixed delta url adjust
Delta have special handling of url which using info fetched from one
page and then fetching the proper menu page. The text aroung the menu
link on the first page now has the year as well as the week info. The
year is not important so the regex was simply updated to gobble it.