Cleanup of html tags for Medicon
Even though the regex is now taking out what we need instead of removing
what we don't need, there are still inconsistencies that make it hard to
do a clean regex. The remaining tags are not shown in the html
rendering, but is part of the source and shows in the json data.

To be extra sure, tags are also removed when doing json convertion.