Fixed weeknum test
The weeknum test on the form 'v.1' for week 1 is to loose and will match
any reference of version 1 in page. This is specifically a problem for
Bryggan. By adding that the week number may not be followed by a dot,
the problem for Bryggan is fixed. Also, adding back weeknum test for
Finn Inn. It is time for them to put the correct week number on their
page.